requirements.txt
accelerate==0.33.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
boto3==1.34.150
botocore==1.34.150
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.6.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastavro==1.9.5
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.3
idna==3.7
ImageHash==4.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
lida==0.0.14
llmx==0.0.21a0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-venn==1.0.0
mdurl==0.1.2
mizani==0.11.4
mpmath==1.3.0
multimethod==1.9.1
networkx==3.3
numba==0.60.0
numpy==1.23.5
openai==1.37.1
packaging==24.1
pandas==1.5.3
pandas-profiling==3.6.3
parameterized==0.9.0
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
pipdeptree==2.23.1
plotly==5.23.0
plotnine==0.13.6
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic-settings==2.3.4
pydantic_core==2.18.1
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.28.2
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scipy==1.9.3
seaborn==0.12.2
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.13.5
streamlit==1.37.0
streamlit-pandas-profiling==0.1.3
sympy==1.13.1
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.64.1
transformers==4.43.3
typeguard==2.13.3
typer==0.12.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.3
visions==0.7.5
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
wordcloud==1.9.3
ydata-profiling==4.9.0
